+++ internal/infrastructure/kubernetes/proxy_infra_test.go
