+++ internal/infrastructure/kubernetes/proxy_serviceaccount_test.go
@@ -213,15 +213,12 @@ func TestCreateOrUpdateProxyServiceAccount(t *testing.T) {
 
 func TestDeleteProxyServiceAccount(t *testing.T) {
 	testCases := []struct {
-		name   string
-		expect bool
+		name string
 	}{
 		{
-			name:   "delete service account",
-			expect: false,
+			name: "delete service account",
 		},
 	}
-
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
@@ -230,7 +227,14 @@ func TestDeleteProxyServiceAccount(t *testing.T) {
 				Namespace: "test",
 			}
 			infra := ir.NewInfra()
-			err := kube.deleteProxyServiceAccount(context.Background(), infra)
+
+			infra.Proxy.GetProxyMetadata().Labels[gatewayapi.OwningGatewayNamespaceLabel] = "default"
+			infra.Proxy.GetProxyMetadata().Labels[gatewayapi.OwningGatewayNameLabel] = infra.Proxy.Name
+
+			err := kube.createOrUpdateProxyServiceAccount(context.Background(), infra)
+			require.NoError(t, err)
+
+			err = kube.deleteProxyServiceAccount(context.Background(), infra)
 			require.NoError(t, err)
 		})
 	}
