+++ internal/infrastructure/kubernetes/ratelimit_labels.go
@@ -1,0 +1,13 @@
+// Copyright Envoy Gateway Authors
+// SPDX-License-Identifier: Apache-2.0
+// The full text of the Apache license is available in the LICENSE file at
+// the root of the repo.
+
+package kubernetes
+
+// rateLimitLabels returns the labels used for all envoy rate limit resources.
+func rateLimitLabels() map[string]string {
+	return map[string]string{
+		"app.gateway.envoyproxy.io/name": rateLimitInfraName,
+	}
+}
